refactor: move toaster to BasePageObject class
